feat: add CddaConfig system for flexible data path configuration
Add cdda.config.json and CddaConfig.ts to support running from any location
by reading data paths from a configuration file. This allows the project to
work with Cataclysm-DDA data from different directory structures.

Changes:
- Add cdda.config.json for centralized path configuration
- Add CddaConfig.ts with auto-discovery of project root and config loading
- Update all CLI tools to use the new config system
- Add test scripts for palette loading and mapgen validation
- Remove obsolete verify-data.cjs and check-tests.sh files
- Update test files to use new configuration approach

+++ packages/core/src/map/MapTileSoa.ts
@@ -11,7 +11,7 @@ import { TrapId } from '../trap/types';
  */
 export class MapTileSoa {
   readonly terrain: Uint16Array;
-  readonly furniture: Uint16Array;
+  readonly furniture: Uint32Array; // 使用 Uint32Array 以支持更大的家具 ID
   readonly radiation: Uint16Array;
   readonly traps: Map<string, TrapId>; // 使用 Map 存储字符串 trap ID
   readonly fields: Map<string, FieldEntry>;
@@ -20,7 +20,7 @@ export class MapTileSoa {
     const count = size * size;
 
     this.terrain = new Uint16Array(count);
-    this.furniture = new Uint16Array(count);
+    this.furniture = new Uint32Array(count);
     this.radiation = new Uint16Array(count);
     this.traps = new Map();
     this.fields = new Map();
